+++ js/data.js
@@ -1,0 +1,30 @@
+const userNames =
+  [
+    'Артём',
+    'Кекс',
+    'Саша',
+    'Юля',
+    'Лена',
+  ];
+
+const commentsMessage =
+  [
+    'Всё отлично!',
+    'В целом всё неплохо. Но не всё.',
+    'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+    'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+    'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+    'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+  ];
+
+
+const photoDescriprion =
+  [
+    'Закат на берегу океана',
+    'Уютный кофейный уголок с книгой',
+    'Радостный день на природе с семьей',
+    'Городской пейзаж с высоты птичьего полета',
+    'Восход солнца в горах'
+  ];
+
+export { userNames, commentsMessage, photoDescriprion }
